finalizando  cap 7, 7.3 fazendo ajuste para passar a imageUrl e configurando imageBinding

+++ app/src/main/java/com/example/myapplication/model/domain/Creature.kt
@@ -2,7 +2,7 @@ package com.example.myapplication.model.domain
 
 
 data class Creature(
-    val number: Long,
+    val number: Int,
     val name: String,
     val imageUrl: String
 )
